moved everything but the main function to the lib
removed [[bin]] sections in Cargo.toml, because Cargo autodetects
binaries inside the bin folder
